% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbo_functions.R
\name{capEx}
\alias{capEx}
\title{Capital Expenditures}
\usage{
capEx(revenue, rate)
}
\arguments{
\item{revenue}{vector of projected revenue for each year post-transaction}

\item{rate}{Cap Ex investment rate (\% of sales) ex: 2\% would be .02}
}
\description{
This function calculates the CapEx given projected revenue and
}
\examples{
capEx(c(1610.51,	1771.561,	1948.7171,	2143.58881,	2357.947691), .02)
}
