% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbo_functions.R
\name{function1}
\alias{function1}
\title{A basic function}
\usage{
function1(x, y)
}
\arguments{
\item{x}{first number to add}

\item{y}{second number to add}
}
\description{
This function allows you to add two number together
}
\examples{
function1(1, 2)
}
\keyword{add,}
\keyword{sum}
