% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactionmultiple.R
\name{transaction_multiple}
\alias{transaction_multiple}
\title{Transaction Multiple Estimator}
\usage{
transaction_multiple(sector, performance, RoIC, risk)
}
\arguments{
\item{sector}{Which engineering sector is your company a part of? View list of current
supported sectors at https://github.com/macbarnes04/lbomodel}

\item{performance}{On a percentile scale (1-99) compared to other companies you have seen, what is you expectation of this company meeting future performance expectations (99\% meaning it would meet expectations better than 99\% of current peer companies, esentially equating to a 99\% confidence in meeting expectations). input as 0-1 (ex. 0.99 means 99 percentile)}

\item{RoIC}{On a percentile scale (1-99) how well has this company returned invested capital relative to its peers? (99 meaning it has a RoIC better than 99\% of peers you have seen). Input as 0-1 (ex. 0.99 means 99 percentile)}

\item{risk}{On a percentile scale (1-99) what is your perceived risk of this company relative to its peers? input as 0-1 (ex. 0.99 means 99 percentile).}
}
\description{
This function allows you to estimate a transaction multiple for a given company.
}
\examples{
transaction_multiple(aerospace)
}
